core-to-plc: all datatypes are abstract
We make `Unit` and `Bool` work as abstract types. This means we need to
adapt the outputs of our primitive functions.

In practice this means we
need to bind the primitives with new signatures that correspond to the
signatures of therr Haskell representatives, and then wire them up to
the real primitives in their definitions.

This causes a *lot* of diff and makes codegen *awful*, because we're now
declaring every primitive in every program, whether or not it uses it.
However, we opt to be stupid for now, and do some dead-code elimination
later when we go to PIR.

We are still not entirely clear of some types needing to be transparent:
`Int` and `ByteString` need to go exactly to their corresponding Plutus
types, otherwise we'd have to mangle far more of the arguments. But this
is okay, since they're not *data* types, and nobody will do anything
with them other than use builtins, so it's okay to just represent them
as the PLC types directly.

This will be a pain when going to PIR, since we don't have transparent
type aliases, but we can perhaps add those.