SCP-2133: use exceptions in the CEK machine
This refactors the CEK machine to use exceptions instead of `ExceptT`.

This results in a substantial (30-50%) speedup on the validation
benchmarks. (Will post benchmark results in a comment.)

The cost is some evil in `runCekM` which we should inspect *very*
carefully. We also need some more type applications since we no longer
have the nice `MonadError` functional depenency that fixes the error
type.

So far I *haven't* pushed this down into the constant application
machinery. The main reason for this is that we have some awkwardness at
the interface of `applyTypeSchemed` where we need to change the `cause`
part of the error from a value into a term, and I don't want to have to
catch and rethrow an exception there if we can avoid it. It's also
convenient for other users of the CAM to have normal errors there. So
I'm not sure if we *should* change that.